pyproject.toml
[project]
name = "ragagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "docx>=0.2.4",
    "fastapi>=0.115.8",
    "keybert>=0.9.0",
    "langchain-community>=0.3.18",
    "langchain-huggingface>=0.1.2",
    "langchain-postgres>=0.0.13",
    "langchain-text-splitters>=0.3.6",
    "pip>=25.0.1",
    "psycopg>=3.2.5",
    "psycopg-pool>=3.2.5",
    "pydantic>=2.10.6",
    "pypdf2>=3.0.1",
    "python-json-logger>=3.2.1",
    "python-multipart>=0.0.20",
    "rq>=2.1.0",
    "spacy>=3.8.4",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "unstructured>=0.16.23",
    "uvicorn>=0.34.0",
    "yoyo-migrations>=9.0.0",
]
